fix: resolve productivity stats unmarshaling and improve visibility
- Change Karma type to float64 to handle API float values
- Move goals display to the right side of the status bar
- Add persistent error reporting for stats fetch failures

+++ internal/api/types.go
@@ -439,7 +439,7 @@ type ProductivityStats struct {
 	Goals      ProductivityGoals `json:"goals"`
 	DaysItems  []DayItems        `json:"days_items"`
 	WeekItems  []WeekItems       `json:"week_items"`
-	Karma      int               `json:"karma"`
+	Karma      float64           `json:"karma"`
 	KarmaTrend string            `json:"karma_trend"` // "up", "down", or ""
 }
